Move node detail viewmodel
